Refactor preview seek to accept Float values for improved precision

+++ app/src/main/java/app/gyrolet/mpvrx/ui/player/controls/AudioPlayerControls.kt
@@ -1215,7 +1215,7 @@ fun AudioPlayerControls(
         position = currentPosSec,
         committedPosition = currentPosSec,
         duration = currentDurSec.coerceAtLeast(1f),
-        onValueChange = { value -> viewModel.previewSeekTo(value.toInt()) },
+        onValueChange = { value -> viewModel.previewSeekTo(value) },
         onValueChangeFinished = { targetPosition -> viewModel.seekTo(targetPosition.toInt(), fast = false) },
         timersInverted = Pair(false, invertDuration),
         durationTimerOnCLick = { playerPreferences.invertDuration.set(!invertDuration) },
